Resolve dependencies for generator-created cores
Currently, generated cores (i.e. core files written by a `generator`)
are inserted into the list of cores, but their dependencies are ignored.
After this commit, cores are inserted into the list of available cores,
and their dependencies are evaluated as with normal cores.

To do that I had to rework the Edalizer class interface a bit and clean
it up significantly in the process. It is now Python 3-only.

TODO: Tests don't pass at this point, they need to be reworked to pass
a Coremanager instance to Edalizer.