Remove variable expansion from generate names
The current code expands environment variables in
`targets.*.generate` blocks. This behavior is not documented and was
likely introduced in error in commit 4d02c089. Since we haven't released
this feature yet, let's remove it.

Example: before this patch, the following was possible, assuming
`TESTGEN_ENV_VAR` was set to `testgen`.

```
name : ::expandvars:0

generators:
  generator1:
    interpreter: python3
    command: testgen.py

targets:
  default:
    generate:
      - $TESTGEN_ENV_VAR
    toplevel: na

generate:
  testgen:
    generator: generator1
```